add ThemeProviderComponent
