+++ src/context/ThemeConext/index.ts
@@ -1,4 +1,6 @@
-export { default as ThemeProvider } from './ThemeContext';
+export { ThemeProvider } from './ThemeContext';
+
+export { default as ThemeProviderComponent } from './ThemeContext';
 
 export { default as useThemeContext } from './useThemeContext';
